feat(arkworks): try exporting tests
