+++ crates/arkworks/Cargo.toml
@@ -16,6 +16,10 @@ ark-ff = { git = "https://github.com/arkworks-rs/algebra", default-features = fa
 ark-ec = { git = "https://github.com/arkworks-rs/algebra", default-features = false }
 ark-std = { git = "https://github.com/arkworks-rs/utils", default-features = false }
 ark-serialize = { git = "https://github.com/arkworks-rs/algebra", default-features = false }
-rustc-hex = { version = "2", default-features = false }
 
+rustc-hex = { version = "2", default-features = false }
 rand = { version = "0.7", default-features = false }
+
+[features]
+default = []
+tests = []
