+++ crates/arkworks/src/tests/mod.rs
@@ -1,0 +1,7 @@
+//! Tests as exports
+#![cfg(any(test, feature = "tests"))]
+mod bls12_377;
+mod bls12_381;
+mod bn254;
+mod bw6_761;
+mod cp6_782;
